requirements.txt
paho-mqtt
rpi_ws281x
flask